requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
cryptography==44.0.1
dataclasses-json==0.6.5
debugpy==1.8.11
decorator==5.2.1
distro==1.9.0
dnspython==2.4.2
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
jsonpatch==1.33
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-google-genai==2.0.10
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.39
markdown-it-py==2.2.0
MarkupSafe==3.0.2
marshmallow==3.19.0
matplotlib-inline==0.1.7
mdurl==0.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numpy==1.26.4
openai==1.78.0
orjson==3.10.14
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==25.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.6.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.11.3
rsa==4.7.2
setuptools==78.1.1
shellingham==1.5.0
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.39
stack_data==0.6.3
starlette==0.46.2
tenacity==9.0.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typer-slim==0.15.3
typing_extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==10.4
wheel==0.45.1
yarl==1.18.0
zipp==3.21.0
zstandard==0.23.0
